+++ rasa_core/events/__init__.py
@@ -12,7 +12,6 @@ import typing
 from builtins import str
 
 from rasa_core import utils
-from rasa_core.actions.action import ACTION_LISTEN_NAME
 
 if typing.TYPE_CHECKING:
     from rasa_core.trackers import DialogueStateTracker
@@ -27,7 +26,7 @@ class Event(object):
     - the topic has been set
     - the bot has taken an action
 
-    Events are logged by the Tracker's log_event method.
+    Events are logged by the Tracker's update method.
     This updates the list of turns so that the current state
     can be recovered by consuming the list of turns."""
 
@@ -55,6 +54,9 @@ class Event(object):
     def _from_story_string(cls, event_name, parameters, domain):
         """Called to convert a parsed story line into an event."""
         return cls()
+
+    def as_dict(self):
+        return {"event": self.type_name}
 
     @classmethod
     def _from_parameters(cls, event_name, parameters, domain):
@@ -131,16 +133,18 @@ class UserUttered(Event):
     def empty():
         return UserUttered(None)
 
+    def as_dict(self):
+        raise NotImplementedError()
+
     def as_story_string(self):
         if self.intent:
             if self.entities:
-                entity_strs = ['{}={}'.format(ent['entity'], ent['value'])
-                               for ent in self.entities]
-                ent_string = "[" + ",".join(entity_strs) + "]"
+                ent_string = json.dumps({ent['entity']: ent['value']
+                                         for ent in self.entities})
             else:
                 ent_string = ""
 
-            return "_{intent}{entities}".format(
+            return "{intent}{entities}".format(
                     intent=self.intent.get("name", ""),
                     entities=ent_string)
         else:
@@ -150,6 +154,57 @@ class UserUttered(Event):
         # type: (DialogueStateTracker) -> None
 
         tracker.latest_message = self
+
+
+# noinspection PyProtectedMember
+class BotUttered(Event):
+    """The bot has said something to the user.
+
+    This class is not used in the story training as it is contained in the
+
+    ``ActionExecuted`` class. An entry is made in the ``Tracker``."""
+
+    type_name = "bot"
+
+    def __init__(self, text=None, data=None):
+        self.text = text
+        self.data = data
+
+    def __hash__(self):
+        return hash((self.text, self.data))
+
+    def __eq__(self, other):
+        if not isinstance(other, BotUttered):
+            return False
+        else:
+            return (self.text, self.data) == \
+                   (other.text, other.data)
+
+    def __str__(self):
+        return "BotUttered(text: {}, data: {})".format(self.text, json.dumps(self.data, indent=2))
+
+    def apply_to(self, tracker):
+        # type: (DialogueStateTracker) -> None
+
+        tracker.latest_bot_utterance = self
+
+    def as_story_string(self):
+        return None
+
+    @staticmethod
+    def empty():
+        return BotUttered()
+
+    def as_dict(self):
+        return {"event": self.type_name,
+                "text": self.text,
+                "data": self.data}
+    @classmethod
+    def _from_parameters(cls, event_name, parameters, domain):
+        try:
+            return BotUttered(parameters["text"], parameters["data"])
+        except KeyError as e:
+            raise ValueError("Failed to parse bot uttered event. {}".format(e))
 
 
 # noinspection PyProtectedMember
@@ -182,6 +237,10 @@ class TopicSet(Event):
     def _from_story_string(cls, event_name, parameters, domain):
         topic = list(parameters.keys())[0] if parameters else ""
         return TopicSet(topic)
+
+    def as_dict(self):
+        return {"event": self.type_name,
+                "topic": self.topic}
 
     @classmethod
     def _from_parameters(cls, event_name, parameters, domain):
@@ -232,6 +291,11 @@ class SlotSet(Event):
         else:
             return None
 
+    def as_dict(self):
+        return {"event": self.type_name,
+                "name": self.key,
+                "value": self.value}
+
     @classmethod
     def _from_parameters(cls, event_name, parameters, domain):
         try:
@@ -264,9 +328,11 @@ class Restarted(Event):
         return self.type_name
 
     def apply_to(self, tracker):
+        from rasa_core.actions.action import ActionListen
         tracker._reset()
-        tracker.update(ActionExecuted(ACTION_LISTEN_NAME))
-        tracker.latest_restart_event = len(tracker.events)
+        # will be the index of the first event after the restart
+        tracker.latest_restart_event = len(tracker.events) + 1
+        tracker.follow_up_action = ActionListen()
 
 
 # noinspection PyProtectedMember
@@ -369,6 +435,9 @@ class ReminderScheduled(Event):
             "name": self.name,
             "kill_on_user_msg": self.kill_on_user_message})
         return "{name}{props}".format(name=self.type_name, props=props)
+
+    def as_dict(self):
+        raise NotImplementedError()
 
     @classmethod
     def _from_story_string(cls, event_name, parameters, domain):