+++ rasa_core/processor.py
@@ -17,9 +17,12 @@ from rasa_core.channels import UserMessage, InputChannel
 from rasa_core.channels.direct import CollectingOutputChannel
 from rasa_core.dispatcher import Dispatcher
 from rasa_core.domain import Domain
-from rasa_core.events import Restarted, ReminderScheduled, Event, SlotSet
-from rasa_core.events import UserUttered, ActionExecuted
-from rasa_core.interpreter import NaturalLanguageInterpreter
+from rasa_core.events import ReminderScheduled, Event
+from rasa_core.events import SlotSet
+from rasa_core.events import UserUttered, ActionExecuted, BotUttered
+from rasa_core.interpreter import (
+    NaturalLanguageInterpreter,
+    INTENT_MESSAGE_PREFIX)
 from rasa_core.interpreter import RegexInterpreter
 from rasa_core.policies.ensemble import PolicyEnsemble
 from rasa_core.tracker_store import TrackerStore
@@ -117,7 +120,7 @@ class MessageProcessor(object):
         tracker = self._get_tracker(sender_id)
         if executed_action != ACTION_LISTEN_NAME:
             self._log_action_on_tracker(tracker, executed_action, events)
-        if self._should_predict_another_action(executed_action, events):
+        if self.should_predict_another_action(executed_action, events):
             return self._predict_next_and_return_state(tracker)
         else:
             self._save_tracker(tracker)
@@ -187,15 +190,22 @@ class MessageProcessor(object):
         # for testing - you can short-cut the NLU part with a message
         # in the format _intent[entity1=val1,entity=val2]
         # parse_data is a dict of intent & entities
-        if message.text.startswith('_'):
+        if (message.text.startswith(INTENT_MESSAGE_PREFIX) or
+                message.text.startswith("_")):
+            if RegexInterpreter.is_using_deprecated_format(message.text):
+                warnings.warn(
+                        "Parsing messages with leading `_` is deprecated and "
+                        "will be removed. Instead, prepend your intents with "
+                        "`{0}`, e.g. `{0}mood_greet`  "
+                        "or `{0}restart`.".format(INTENT_MESSAGE_PREFIX))
             parse_data = RegexInterpreter().parse(message.text)
         else:
             parse_data = self.interpreter.parse(message.text)
 
         logger.debug("Received user message '{}' with intent '{}' "
-                     "and entities  '{}'".format(message.text,
-                                                 parse_data["intent"],
-                                                 parse_data["entities"]))
+                     "and entities '{}'".format(message.text,
+                                                parse_data["intent"],
+                                                parse_data["entities"]))
         return parse_data
 
     def _handle_message_with_tracker(self, message, tracker):
@@ -203,7 +213,7 @@ class MessageProcessor(object):
 
         parse_data = self._parse_message(message)
 
-        # don't ever directly mutate the tracker - instead pass it events to log
+        # don't ever directly mutate the tracker - instead pass its events to log
         tracker.update(UserUttered(message.text, parse_data["intent"],
                                    parse_data["entities"], parse_data))
         # store all entities as slots
@@ -231,9 +241,9 @@ class MessageProcessor(object):
         self._log_slots(tracker)
 
         # action loop. predicts actions until we hit action listen
-        while should_predict_another_action and \
-                self._should_handle_message(tracker) and \
-                        num_predicted_actions < self.max_number_of_predictions:
+        while (should_predict_another_action
+               and self._should_handle_message(tracker)
+               and num_predicted_actions < self.max_number_of_predictions):
             # this actually just calls the policy's method by the same name
             action = self._get_next_action(tracker)
 
@@ -242,8 +252,8 @@ class MessageProcessor(object):
                                                              dispatcher)
             num_predicted_actions += 1
 
-        if num_predicted_actions == self.max_number_of_predictions and \
-                should_predict_another_action:
+        if (num_predicted_actions == self.max_number_of_predictions and
+                should_predict_another_action):
             # circuit breaker was tripped
             logger.warn(
                     "Circuit breaker tripped. Stopped predicting "
@@ -254,10 +264,10 @@ class MessageProcessor(object):
 
         logger.debug("Current topic: {}".format(tracker.topic.name))
 
-    def _should_predict_another_action(self, action_name, events):
+    @staticmethod
+    def should_predict_another_action(action_name, events):
         is_listen_action = action_name == ACTION_LISTEN_NAME
-        contains_restart = events and isinstance(events[0], Restarted)
-        return not is_listen_action and not contains_restart
+        return not is_listen_action
 
     def _schedule_reminders(self, events, dispatcher):
         # type: (List[Event], Dispatcher) -> None
@@ -287,10 +297,11 @@ class MessageProcessor(object):
                          "code.".format(action.name()), )
             logger.error(e, exc_info=True)
             events = []
+        self.log_bot_utterances_on_tracker(tracker, dispatcher)
         self._log_action_on_tracker(tracker, action.name(), events)
         self._schedule_reminders(events, dispatcher)
 
-        return self._should_predict_another_action(action.name(), events)
+        return self.should_predict_another_action(action.name(), events)
 
     def _warn_about_new_slots(self, tracker, action_name, events):
         # these are the events from that action we have seen during training
@@ -309,6 +320,18 @@ class MessageProcessor(object):
                                 "adding '- slot{{\"{1}\": \"{2}\"}}' "
                                 "after the action."
                                 "".format(action_name, e.key, e.value))
+
+    @staticmethod
+    def log_bot_utterances_on_tracker(tracker, dispatcher):
+        # type: (DialogueStateTracker, Dispatcher) -> None
+
+        if dispatcher.latest_bot_messages:
+            for m in dispatcher.latest_bot_messages:
+                bot_utterance = BotUttered(text=m.text, data=m.data)
+                logger.debug("Bot utterance '{}'".format(bot_utterance))
+                tracker.update(bot_utterance)
+
+            dispatcher.latest_bot_messages = []
 
     def _log_action_on_tracker(self, tracker, action_name, events):
         # Ensures that the code still works even if a lazy programmer missed
@@ -354,8 +377,8 @@ class MessageProcessor(object):
                         "Instead of running that, we will ignore the action "
                         "and predict the next action.".format(follow_up_action))
 
-        if tracker.latest_message.intent.get("name") == \
-                self.domain.restart_intent:
+        if (tracker.latest_message.intent.get("name") ==
+                self.domain.restart_intent):
             return ActionRestart()
 
         idx = self.policy_ensemble.predict_next_action(tracker, self.domain)