+++ tests/test_domain.py
@@ -10,13 +10,13 @@ from rasa_core.domain import TemplateDomain
 from rasa_core.featurizers import BinaryFeaturizer
 from rasa_core.training import extract_training_data_from_file
 from tests import utilities
-from tests.conftest import DEFAULT_DOMAIN_PATH
+from tests.conftest import DEFAULT_DOMAIN_PATH, DEFAULT_STORIES_FILE
 
 
 def test_create_train_data_no_history(default_domain):
     featurizer = BinaryFeaturizer()
     training_data = extract_training_data_from_file(
-            "data/dsl_stories/stories_defaultdomain.md",
+            DEFAULT_STORIES_FILE,
             default_domain,
             featurizer,
             augmentation_factor=0,
@@ -47,7 +47,7 @@ def test_create_train_data_no_history(default_domain):
 def test_create_train_data_with_history(default_domain):
     featurizer = BinaryFeaturizer()
     training_data = extract_training_data_from_file(
-            "data/dsl_stories/stories_defaultdomain.md",
+            DEFAULT_STORIES_FILE,
             default_domain,
             featurizer,
             augmentation_factor=0,