+++ tests/test_dsl.py
@@ -14,7 +14,7 @@ from rasa_core.training.structures import Story
 
 
 def test_can_read_test_story(default_domain):
-    trackers = extract_trackers_from_file("data/dsl_stories/stories.md",
+    trackers = extract_trackers_from_file("data/test_stories/stories.md",
                                           default_domain,
                                           featurizer=BinaryFeaturizer())
     assert len(trackers) == 7
@@ -37,7 +37,7 @@ def test_can_read_test_story(default_domain):
 
 
 def test_persist_and_read_test_story_graph(tmpdir, default_domain):
-    graph = extract_story_graph_from_file("data/dsl_stories/stories.md",
+    graph = extract_story_graph_from_file("data/test_stories/stories.md",
                                           default_domain)
     out_path = tmpdir.join("persisted_story.md")
     with io.open(out_path.strpath, "w") as f:
@@ -59,7 +59,7 @@ def test_persist_and_read_test_story_graph(tmpdir, default_domain):
 
 
 def test_persist_and_read_test_story(tmpdir, default_domain):
-    graph = extract_story_graph_from_file("data/dsl_stories/stories.md",
+    graph = extract_story_graph_from_file("data/test_stories/stories.md",
                                           default_domain)
     out_path = tmpdir.join("persisted_story.md")
     Story(graph.story_steps).dump_to_file(out_path.strpath)