+++ tests/test_policies.py
@@ -21,7 +21,7 @@ from tests.conftest import DEFAULT_DOMAIN_PATH, DEFAULT_STORIES_FILE
 
 def train_data(max_history, domain):
     return extract_training_data_from_file(
-            "data/dsl_stories/stories_defaultdomain.md", domain,
+            DEFAULT_STORIES_FILE, domain,
             BinaryFeaturizer(), max_history=max_history, remove_duplicates=True)