+++ tests/test_trackers.py
@@ -64,7 +64,7 @@ def test_tracker_store_storage_and_retrieval(store):
 
     # lets log a test message
     intent = {"name": "greet", "confidence": 1.0}
-    tracker.update(UserUttered("_greet", intent, []))
+    tracker.update(UserUttered("/greet", intent, []))
     assert tracker.latest_message.intent.get("name") == "greet"
     store.save(tracker)
 
@@ -103,19 +103,33 @@ def test_tracker_write_to_story(tmpdir, default_domain):
     assert recovered.events[6] == SlotSet("location", "central")
 
 
-def test_tracker_state_regression(default_agent):
-    sender_id = "test_tracker_state_regression"
-    n_actions = []
+def test_tracker_state_regression_without_bot_utterance(default_agent):
+    sender_id = "test_tracker_state_regression_without_bot_utterance"
     for i in range(0, 2):
-        default_agent.handle_message("_greet", sender_id=sender_id)
+        default_agent.handle_message("/greet", sender_id=sender_id)
     tracker = default_agent.tracker_store.get_or_create_tracker(sender_id)
 
     # Ensures that the tracker has changed between the utterances
     # (and wasn't reset in between them)
     expected = ("action_listen;"
-                "_greet;utter_greet;action_listen;"
-                "_greet;action_listen")
-    assert ";".join([e.as_story_string() for e in tracker.events]) == expected
+                "greet;utter_greet;action_listen;"
+                "greet;action_listen")
+    assert ";".join([e.as_story_string() for e in
+                     tracker.events if e.as_story_string()]) == expected
+
+
+def test_tracker_state_regression_with_bot_utterance(default_agent):
+    sender_id = "test_tracker_state_regression_with_bot_utterance"
+    for i in range(0, 2):
+        default_agent.handle_message("/greet", sender_id=sender_id)
+    tracker = default_agent.tracker_store.get_or_create_tracker(sender_id)
+
+    expected = ["action_listen", "greet", None, "utter_greet",
+                "action_listen", "greet", "action_listen"]
+    print([e.as_story_string() for e in tracker.events])
+    for e in tracker.events:
+        print(e)
+    assert [e.as_story_string() for e in tracker.events] == expected
 
 
 def test_tracker_entity_retrieval(default_domain):
@@ -127,7 +141,7 @@ def test_tracker_entity_retrieval(default_domain):
     assert list(tracker.get_latest_entity_values("entity_name")) == []
 
     intent = {"name": "greet", "confidence": 1.0}
-    tracker.update(UserUttered("_greet", intent, [{
+    tracker.update(UserUttered("/greet", intent, [{
           "start": 1,
           "end": 5,
           "value": "greet",
@@ -147,19 +161,21 @@ def test_restart_event(default_domain):
 
     intent = {"name": "greet", "confidence": 1.0}
     tracker.update(ActionExecuted(ACTION_LISTEN_NAME))
-    tracker.update(UserUttered("_greet", intent, []))
+    tracker.update(UserUttered("/greet", intent, []))
     tracker.update(ActionExecuted("my_action"))
     tracker.update(ActionExecuted(ACTION_LISTEN_NAME))
 
     assert len(tracker.events) == 4
-    assert tracker.latest_message.text == "_greet"
+    assert tracker.latest_message.text == "/greet"
     assert len(list(tracker.generate_all_prior_states())) == 4
 
     tracker.update(Restarted())
 
-    assert len(tracker.events) == 6
+    assert len(tracker.events) == 5
+    assert tracker.follow_up_action is not None
+    assert tracker.follow_up_action.name() == ACTION_LISTEN_NAME
     assert tracker.latest_message.text is None
-    assert len(list(tracker.generate_all_prior_states())) == 2
+    assert len(list(tracker.generate_all_prior_states())) == 1
 
     dialogue = tracker.as_dialogue()
 
@@ -169,9 +185,11 @@ def test_restart_event(default_domain):
     recovered.recreate_from_dialogue(dialogue)
 
     assert recovered.current_state() == tracker.current_state()
-    assert len(recovered.events) == 6
+    assert len(recovered.events) == 5
+    assert tracker.follow_up_action is not None
+    assert tracker.follow_up_action.name() == ACTION_LISTEN_NAME
     assert recovered.latest_message.text is None
-    assert len(list(recovered.generate_all_prior_states())) == 2
+    assert len(list(recovered.generate_all_prior_states())) == 1
 
 
 def test_revert_action_event(default_domain):
@@ -183,7 +201,7 @@ def test_revert_action_event(default_domain):
 
     intent = {"name": "greet", "confidence": 1.0}
     tracker.update(ActionExecuted(ACTION_LISTEN_NAME))
-    tracker.update(UserUttered("_greet", intent, []))
+    tracker.update(UserUttered("/greet", intent, []))
     tracker.update(ActionExecuted("my_action"))
     tracker.update(ActionExecuted(ACTION_LISTEN_NAME))